+++ pkg/vif/device_linux.go
@@ -12,7 +12,6 @@ import (
 
 	"github.com/datawire/dlib/dexec"
 	"github.com/telepresenceio/telepresence/v2/pkg/vif/buffer"
-	"github.com/telepresenceio/telepresence/v2/pkg/vif/routing"
 )
 
 const devicePath = "/dev/net/tun"
@@ -98,14 +97,6 @@ func (t *Device) addSubnet(ctx context.Context, subnet *net.IPNet) error {
 
 func (t *Device) removeSubnet(ctx context.Context, subnet *net.IPNet) error {
 	return dexec.CommandContext(ctx, "ip", "a", "del", subnet.String(), "dev", t.name).Run()
-}
-
-func (t *Device) addStaticRoute(ctx context.Context, route *routing.Route) error {
-	return dexec.CommandContext(ctx, "ip", "route", "add", route.RoutedNet.String(), "via", route.Gateway.String(), "dev", route.Interface.Name).Run()
-}
-
-func (t *Device) removeStaticRoute(ctx context.Context, route *routing.Route) error {
-	return dexec.CommandContext(ctx, "ip", "route", "del", route.RoutedNet.String(), "via", route.Gateway.String(), "dev", route.Interface.Name).Run()
 }
 
 // Index returns the index of this device
