+++ pkg/vif/routing/routing.go
@@ -5,6 +5,8 @@ import (
 	"errors"
 	"fmt"
 	"net"
+
+	"github.com/datawire/dlib/dlog"
 )
 
 type Route struct {
@@ -37,6 +39,19 @@ func (r *Route) String() string {
 		return fmt.Sprintf("default via %s dev %s, gw %s", r.LocalIP, r.Interface.Name, r.Gateway)
 	}
 	return fmt.Sprintf("%s via %s dev %s, gw %s", r.RoutedNet, r.LocalIP, r.Interface.Name, r.Gateway)
+}
+
+// AddStatic adds a specific route. This can be used to prevent certain IP addresses
+// from being routed to the route's interface.
+func (r *Route) AddStatic(ctx context.Context) error {
+	dlog.Debugf(ctx, "Adding static route %s", r)
+	return r.addStatic(ctx)
+}
+
+// RemoveStatic removes a specific route added via AddStatic.
+func (r *Route) RemoveStatic(ctx context.Context) error {
+	dlog.Debugf(ctx, "Dropping static route %s", r)
+	return r.removeStatic(ctx)
 }
 
 func interfaceLocalIP(iface *net.Interface, ipv4 bool) (net.IP, error) {
